requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
fastapi[all]
python-dateutil
google-auth 
certifi 
requests 
urllib3
pytz
google-genai
openai
vertexai
pyyaml
google-cloud-pubsub
python-dotenv
colorama
python-jose
sqlalchemy[asyncio]
asyncpg
redis
alembic
psycopg2-binary